-- complex delete

+++ 20250816/hecto/src/edit/view.rs
@@ -101,7 +101,7 @@ impl View {
         let grapheme_delta = new_len.saturating_sub(old_len);
         if grapheme_delta > 0 {
             //move right for an added grapheme (should be the regular case)
-            self.move_right();
+            self.move_text_location(&Direction::Right);
         }
         self.need_redraw = true;
     }
@@ -175,8 +175,10 @@ impl View {
     }
 
     fn backspace(&mut self) {
-        self.move_left();
-        self.delete()
+        if self.text_location.line_index != 0 || self.text_location.grapheme_index != 0 {
+            self.move_text_location(&Direction::Left);
+            self.delete();
+        }
     }
 
     fn delete(&mut self) {
